@@ -12,6 +12,6 @@ namespace Cerberus.Models
         public DateTime? LastUpdateDate { get; set; }
         public int TotalChapters { get; set; }
         public int TotalVolumes { get; set; }
-        public ICollection<Language> UserLanguages { get; set; }
+        public IList<Language> UserLanguages { get; set; }
     }
 }
